app/requirements.txt
torch
torchvision
gradio
Pillow
tqdm
